feat: enhance authentication flow with phone verification and loading indicators; refactor related components
