+++ viewModels/authViewModel.ts
@@ -7,6 +7,9 @@ export const useAuthViewModel = () => {
   const router = useRouter();
   const [loading, setLoading] = useState(false);
   const [isCodeSent, setIsCodeSent] = useState(false);
+  const [showVerification, setShowVerification] = useState(false);
+  const [phoneVerified, setPhoneVerified] = useState(false);
+  const [phone, setPhone] = useState('');
 
   const signInWithEmail = async (email: string, password: string) => {
     setLoading(true);
@@ -24,15 +27,12 @@ export const useAuthViewModel = () => {
     setLoading(false);
   };
 
-  const handlePhoneLogin = async (
-    formattedValue: string,
-    verificationCode: string,
-  ) => {
+  const signInWithPhone = async (verificationCode: string) => {
     setLoading(true);
     try {
       if (!isCodeSent) {
         const { error } = await supabase.auth.signInWithOtp({
-          phone: formattedValue,
+          phone,
         });
 
         if (error) {
@@ -46,7 +46,7 @@ export const useAuthViewModel = () => {
         }
       } else {
         const { error } = await supabase.auth.verifyOtp({
-          phone: formattedValue,
+          phone,
           token: verificationCode,
           type: 'sms',
         });
@@ -66,10 +66,125 @@ export const useAuthViewModel = () => {
     }
   };
 
+  const checkSession = async () => {
+    const {
+      data: { session },
+    } = await supabase.auth.getSession();
+    if (session) {
+      router.replace('/(tabs)/home');
+    }
+  };
+
+  const sendVerificationCode = async () => {
+    if (!phone) {
+      Alert.alert('Error', 'Please enter your phone number');
+      return;
+    }
+    setLoading(true);
+    try {
+      const { error } = await supabase.auth.signUp({
+        phone,
+        password: Math.random().toString(36).slice(-8),
+      });
+
+      if (error) throw error;
+
+      setShowVerification(true);
+      Alert.alert('Success', 'Verification code sent to your phone');
+    } catch (error) {
+      Alert.alert(
+        'Error',
+        error instanceof Error ? error.message : 'An error occurred',
+      );
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const verifyPhone = async (verificationCode: string) => {
+    if (!verificationCode) {
+      Alert.alert('Error', 'Please enter verification code');
+      return;
+    }
+
+    setLoading(true);
+    try {
+      const { error } = await supabase.auth.verifyOtp({
+        phone,
+        token: verificationCode,
+        type: 'sms',
+      });
+
+      if (error) throw error;
+      setPhoneVerified(true);
+      Alert.alert(
+        'Success',
+        'Phone verified successfully! Please complete your registration.',
+      );
+    } catch (error) {
+      Alert.alert(
+        'Error',
+        error instanceof Error ? error.message : 'An error occurred',
+      );
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const completeSignUp = async (
+    email: string,
+    password: string,
+    confirmPassword: string,
+  ) => {
+    if (!email || !password || !confirmPassword) {
+      Alert.alert('Error', 'Please fill in all fields');
+      return;
+    }
+
+    if (password !== confirmPassword) {
+      Alert.alert('Error', 'Passwords do not match');
+      return;
+    }
+
+    setLoading(true);
+    try {
+      const { error } = await supabase.auth.updateUser({
+        email,
+        password,
+        data: {
+          phone_verified: true,
+          phone,
+        },
+      });
+
+      if (error) throw error;
+
+      Alert.alert('Success', 'Account created successfully!');
+      router.push('/(pages)/(authentication)/login');
+    } catch (error) {
+      Alert.alert(
+        'Error',
+        error instanceof Error ? error.message : 'An error occurred',
+      );
+    } finally {
+      setLoading(false);
+    }
+  };
+
   return {
+    // States
+    phone,
     isCodeSent,
     loading,
+    showVerification,
+    phoneVerified,
+    // Methods
     signInWithEmail,
-    handlePhoneLogin,
+    signInWithPhone,
+    checkSession,
+    sendVerificationCode,
+    verifyPhone,
+    completeSignUp,
+    setPhone,
   };
 };
